Refactor fetch email API helper Return an email only Improve error tracing
